Add MainDisplay file to components, have components rendering there instead of in App. Renamed DisplayData folder and file to ManageData. Updated Imports.
