+++ src/store/reducers.ts
@@ -1,6 +1,6 @@
 import { reducer as weatherReducer } from '../Features/Weather/reducer';
 import metricsOptionsReducer from '../Features/SelectMetrics/reducer';
-import dataReducer from '../Features/DisplayData/reducer';
+import dataReducer from '../Features/ManageData/reducer';
 
 export default {
   weather: weatherReducer,
